chore: rename session_v2.py to session.py
What changed and why:
- Rename session_v2.py to session.py

Impact:
- Completes migration from single client session to multiple client
  sessions.
