+++ tests/server/session/test_notification_handling.py
@@ -4,7 +4,7 @@ from conduit.protocol.base import PROTOCOL_VERSION
 from conduit.protocol.common import CancelledNotification, ProgressNotification
 from conduit.protocol.initialization import Implementation, ServerCapabilities
 from conduit.protocol.roots import ListRootsResult, Root, RootsListChangedNotification
-from conduit.server.session_v2 import ServerConfig, ServerSession
+from conduit.server.session import ServerConfig, ServerSession
 
 
 class TestNotificationHandling:
