+++ tests/server/session/test_prompt_handling.py
@@ -19,7 +19,7 @@ from conduit.protocol.prompts import (
     ListPromptsResult,
     PromptMessage,
 )
-from conduit.server.session_v2 import ServerConfig, ServerSession
+from conduit.server.session import ServerConfig, ServerSession
 
 
 class TestPromptHandling:
